# docs/swagger.yaml
basePath: /
definitions:
  domain.User:
    properties:
      allow:
        items:
          type: string
        type: array
      createdDate:
        type: string
      deny:
        items:
          type: string
        type: array
      id:
        type: integer
      roles:
        items:
          type: string
        type: array
      updatedDate:
        type: string
    type: object
  http.CreateUserBody:
    properties:
      allow:
        items:
          type: string
        type: array
      deny:
        items:
          type: string
        type: array
      roles:
        items:
          type: string
        type: array
    type: object
  http.UpdateUserBody:
    properties:
      allow:
        items:
          type: string
        type: array
      deny:
        items:
          type: string
        type: array
      id:
        type: integer
      roles:
        items:
          type: string
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: some desc
  title: template Service API
  version: "1.0"
paths:
  /user:
    get:
      consumes:
      - application/json
      description: get user
      parameters:
      - description: Bearer jwtToken
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
      summary: get user
      tags:
      - user
    post:
      consumes:
      - application/json
      description: create user
      parameters:
      - description: Bearer jwtToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: body params
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.CreateUserBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
      summary: create user
      tags:
      - user
    put:
      consumes:
      - application/json
      description: update user
      parameters:
      - description: Bearer jwtToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: body params
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.UpdateUserBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
      summary: update user
      tags:
      - user
  /user/{id}:
    get:
      consumes:
      - application/json
      description: get user by id
      parameters:
      - description: Bearer jwtToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
      summary: get user by id
      tags:
      - user
schemes:
- http
swagger: "2.0"